pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hoploy"]

[tool.hatch.version]
path = "hoploy/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true


[project]
name = "hoploy"
version = "0.1.0"
description = "Serving layer for Hopwise recommendation models"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110",
    "hydra-core>=1.3",
    "dgl",
    "igraph~=0.10.0",
    "transformers<4.56.0",
    "numpy",
    "omegaconf>=2.3",
    "pydantic>=2.0",
    "uvicorn[standard]>=0.29",
    "faiss-gpu>=1.7.2",
    "hopwise[pathlm, cli] @ git+https://github.com/tail-unica/hopwise.git@24-add-support-for-user-side-kg-and-testing",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
]

[project.scripts]
hoploy = "hoploy.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["hoploy*"]

[tool.uv]
find-links = [
    "https://data.dgl.ai/wheels/torch-2.4/cu124/repo.html"
]
dependency-metadata = [
    { name = "dgl", requires-dist = [], requires-python = ">=3.10" },
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ruff",
    "scalene>=1.5.51",
    "scikit-learn>=1.6.1",
    "scipy>=1.13.1",
    "yellowbrick>=1.5",
]

# --- Ruff Configuration (linting) ---

[tool.ruff]
line-length = 119

[tool.ruff.lint]
select = [
  # Pylint
  "PL",
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # isort
  "I",
  # pyupgrade
  "UP",
  # pydocstyle
  # "D",
]
fixable = ["ALL"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.ruff.lint.pydocstyle]
convention = "google"
